+++ src/Naneau/FileGen/Directory.php
@@ -3,7 +3,7 @@ namespace Naneau\FileGen;
 
 use Naneau\FileGen\Exception as FileGenException;
 
-use \Iterator;
+use Iterator;
 
 /**
  * A directory, that can contain children (other directories, files, symlinks)
@@ -14,17 +14,15 @@ class Directory extends AccessRights implements Iterator
 {
     /**
      * Position of the iteration
-     *
-     * @var int
      */
-    private $position = 0;
+    private int $position = 0;
 
     /**
      * Child nodes
      *
      * @var Node[]
      */
-    private $children = [];
+    private array $children = [];
 
     public function __construct(string $name, int $mode = 0777)
     {
@@ -38,11 +36,9 @@ class Directory extends AccessRights implements Iterator
      * exists, it has a child directory node `bar`, which should have a child
      * node `baz`
      *
-     * Will return either the found child node, or boolean false
-     *
-     * @return Node|bool
+     * Will return either the found child node, or null.
      */
-    public function scan(string $path)
+    public function scan(string $path): ?Node
     {
         // Start scanning at the root (this dir)
         $node = $this;
@@ -52,12 +48,12 @@ class Directory extends AccessRights implements Iterator
 
             // Can't find children if $node is not a directory
             if (!($node instanceof Directory)) {
-                return false;
+                return null;
             }
 
             // If the current node doesn't have the item, $path doesn't exist (fully)
             if (!$node->hasChild($item)) {
-                return false;
+                return null;
             }
 
             // New parent node
@@ -72,7 +68,7 @@ class Directory extends AccessRights implements Iterator
      *
      * @return Node[]
      */
-    public function getChildren()
+    public function getChildren(): array
     {
         return $this->children;
     }
@@ -117,6 +113,8 @@ class Directory extends AccessRights implements Iterator
 
     /**
      * Get a child with name $name
+     *
+     * @throws Exception If the node was not found.
      */
     public function getChild(string $name): Node
     {